Renamed ConsoleInfo to Console and moved all static function from DeveloperConsoleUI to here. Furthermore started using internal access modifiers to make things easier.
