+++ Scripts/Console.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3f5b54d9aedfb22458a16c3f566f0012
+guid: bc045ae8c1cee8445958a5fd42bae724
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
